only do migration if characteristic doenst exist

+++ src/database/migrations/20250528912_create_clinic_table.js
@@ -1,19 +1,23 @@
 exports.up = function(knex) {
-  return knex.schema.createTable('clinics', function(table){
-    table.increments('id').primary();      
-    table.integer('user_id')
-         .unsigned()
-         .notNullable()
-         .references('id')
-         .inTable('users')
-         .onDelete('CASCADE');      
-    table.string('name', 100).notNullable();
-    table.string('cnpj', 14).notNullable().unique();
-    table.jsonb('address').notNullable();
-    table.string('phone').notNullable();
-    table.string('whatsapp').nullable();
-    table.timestamps(true, true);                   
-  })
+  return knex.schema.hasTable('clinics').then(function(exists){
+    if(!exists) {
+      return knex.schema.createTable('clinics', function(table){
+      table.increments('id').primary();      
+      table.integer('user_id')
+           .unsigned()
+           .notNullable()
+           .references('id')
+           .inTable('users')
+           .onDelete('CASCADE');      
+      table.string('name', 100).notNullable();
+      table.string('cnpj', 14).notNullable().unique();
+      table.jsonb('address').notNullable();
+      table.string('phone').notNullable();
+      table.string('whatsapp').nullable();
+      table.timestamps(true, true);                   
+      });
+    }
+  });
 }
 
 exports.down = function(knex) {
